backend/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
bcrypt
pydantic[email]

